refactor(mesh): modularise mesh/manipulate
